Orb Death, More Gnome Lift etc.
Got a long list of things to do, but gonna go for a run to clear my head so I can punch as much of this as I can out today. Need to have time this weekend to prep for work on Monday, so ideally be done Saturday evening or Sunday morning.